Title: 사각형 채우기 3, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240219/사각형 채우기 3/rectangle-fill-3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int dp[1001];
+long long dp[1001];
 
 int main() {
     int N;
